requirements.txt
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.1
django-taggit==1.3.0
pytz==2020.1
sqlparse==0.3.1
